+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## [Unreleased]
 
+### Added
+
+- `YQ` data source.
+
+### Changed
+
+- `JQ` data source `query` arg has been renamed to `expression`.
+
 ## [v0.1.0] - 2022-08-06
 
 ### Added
 
 - Bootstrap provider.
-- JQ data source.
+- `JQ` data source.
 - Released on Terraform registry.
 
 [unreleased]: https://github.com/slok/terraform-provider-dataprocessor/compare/v0.1.0...HEAD